+++ packages/core/src/commands/index.ts
@@ -3,6 +3,7 @@ export * from './intention';
 export * from './select';
 export * from './insert-content';
 export * from './remove-content';
+export * from './append-blocks';
 export * from './change-property';
 export * from './change-title';
 export * from './change-value';