+++ packages/core/src/extensions/row.tsx
@@ -1,28 +1,12 @@
-import { useRef } from 'react';
-import ReactDOM from 'react-dom';
+import { useRef, useState } from 'react';
 import clsx from 'clsx';
 
-import { Button, ButtonGroup } from '@chamaeleon/uikit';
+import { Button, Input } from '@chamaeleon/uikit';
 
 import { BlockExtension } from '../block-extension';
 import { Block } from '../model';
 import { Plugin, PluginKey } from '../state';
 import { JSONContent } from '../types';
-
-const columnMap = {
-  1: 'grid-cols-1',
-  2: 'grid-cols-2',
-  3: 'grid-cols-3',
-  4: 'grid-cols-4',
-  5: 'grid-cols-5',
-  6: 'grid-cols-6',
-  7: 'grid-cols-7',
-  8: 'grid-cols-8',
-  9: 'grid-cols-9',
-  10: 'grid-cols-10',
-  11: 'grid-cols-11',
-  12: 'grid-cols-12',
-} as Record<number, string>;
 
 declare module '..' {
   interface Commands<ReturnType> {
@@ -36,6 +20,65 @@ declare module '..' {
   }
 }
 
+const splitColumns = (template: string) => {
+  return template.split('+').map(Number).filter(Boolean);
+};
+
+const TemplateButton = ({
+  className,
+  template,
+  onClick,
+}: {
+  className?: string;
+  template: string;
+  onClick: (value: number[]) => void;
+}) => {
+  const templateColumns = splitColumns(template);
+
+  if (templateColumns.length === 0) return null;
+
+  return (
+    <button
+      key={template}
+      className={clsx('group flex w-1/5 flex-col items-center p-2', className)}
+      onClick={() => onClick(templateColumns)}
+    >
+      <div className="mb-1 w-full rounded-lg bg-gray-100 p-2 transition group-hover:bg-blue-200 group-hover:outline-none group-hover:ring-1 group-hover:ring-blue-500">
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          className="fill-slate-500 group-hover:fill-blue-700"
+          width="100%"
+          height="25"
+          viewBox="0 0 125 40"
+        >
+          {templateColumns.map((columnSize, index, array) => {
+            const calcX = (index: number): number => {
+              if (index === 0) return 0;
+
+              return array[index - 1] * 10 + calcX(index - 1);
+            };
+
+            const offset = 5;
+
+            return (
+              <rect
+                key={index}
+                width={columnSize * 10 - offset}
+                height="40"
+                x={calcX(index) + offset}
+                fillOpacity={index % 2 === 0 ? '.3' : '.6'}
+                rx="5"
+              />
+            );
+          })}
+        </svg>
+      </div>
+
+      <div className="text-xs group-hover:text-blue-700">{template}</div>
+    </button>
+  );
+};
+
 export const Row = BlockExtension.create({
   name: 'row',
 
@@ -47,18 +90,12 @@ export const Row = BlockExtension.create({
   rootable: true,
   structural: true,
 
-  addProperties() {
-    return {
-      columns: {
-        default: 2,
-        isRequired: true,
-      },
-    };
-  },
-
   addStyle() {
     return {
       root: {
+        gap: {
+          default: '16px',
+        },
         margin: {},
         marginTop: {},
         marginRight: {},
@@ -86,10 +123,7 @@ export const Row = BlockExtension.create({
     return {
       natural: ({ block, children }) => {
         return (
-          <div
-            className={clsx('e-row grid gap-4', columnMap[block.props.columns])}
-            style={block.style.root}
-          >
+          <div className={clsx('e-row grid')} style={block.style.root}>
             {children}
           </div>
         );
@@ -109,7 +143,7 @@ export const Row = BlockExtension.create({
                 <editor.view.Dropzone>
                   <div
                     className={clsx(
-                      'hover:block-highlight flex w-full justify-center bg-white p-5',
+                      'hover:block-highlight flex w-full grid-cols-12 justify-center bg-white p-5',
                       'border-2 border-dashed border-gray-500',
                       {
                         'available-drop': isAvailableDrop,
@@ -133,8 +167,7 @@ export const Row = BlockExtension.create({
             ) : (
               <div
                 className={clsx(
-                  'e-row hover:block-highlight grid w-full gap-4 p-5',
-                  columnMap[block.props.columns],
+                  'e-row hover:block-highlight grid w-full grid-cols-12 p-5',
                   {
                     'available-drop': isAvailableDrop,
                     'dropzone-over': isOver,
@@ -187,47 +220,165 @@ export const Row = BlockExtension.create({
         type: 'property-configuration',
         property: {
           name: 'columns',
+          propertyMatch: false,
           applicable: {
             name: ['row'],
           },
         },
-
         view: (view) => ({
-          update() {
-            if (!view.propertyConfiguration.element) return null;
-
+          render() {
             const { state } = view;
 
-            const handleChangeColumnCount = (count: number) => {
-              if (!state.activeId) return;
+            const [value, setValue] = useState('');
 
-              editor.commands.changeProperty(state.activeId, 'columns', count);
+            const handleChangeTemplate = (columns: number[]) => {
+              const { activeBlock } = state;
+
+              if (!activeBlock) return;
+
+              editor.commands.command(({ commands }) => {
+                if (activeBlock.children.isEmpty) {
+                  columns.forEach((columnSize) => {
+                    commands.addColumn(activeBlock.id, {
+                      colSpan: columnSize,
+                    });
+                  });
+                } else {
+                  columns.forEach((columnSize, index) => {
+                    const childId = activeBlock.children.at(index);
+
+                    if (childId) {
+                      commands.changeProperty(
+                        childId,
+                        'colSpan',
+                        columns[index],
+                      );
+                    } else {
+                      commands.addColumn(activeBlock.id, {
+                        colSpan: columnSize,
+                      });
+                    }
+
+                    const isLast = index === columns.length - 1;
+
+                    if (isLast) {
+                      const restChildren = activeBlock.children.slice(
+                        index + 1,
+                      );
+
+                      if (restChildren.length === 0) return;
+
+                      const childrenFromRemoved = restChildren
+                        .map((child) => {
+                          return state.getBlock(child).children.children;
+                        })
+                        .flat();
+
+                      const futureLastChild = activeBlock.children.at(index);
+
+                      if (!futureLastChild) return;
+
+                      if (childrenFromRemoved.length > 0) {
+                        commands.appendBlocks(
+                          futureLastChild,
+                          childrenFromRemoved,
+                        );
+                      }
+
+                      commands.removeContent(activeBlock.id, restChildren);
+                    }
+                  });
+                }
+              });
             };
 
-            return ReactDOM.createPortal(
-              <div>
-                <p className="text-lg">Select the number of columns</p>
+            const templates = [
+              '12',
+              '6+6',
+              '4+4+4',
+              '3+3+3+3',
+              '4+8',
+              '3+9',
+              '3+6+3',
+              '2+6+4',
+              '2+10',
+              '2+3+7',
+            ];
 
-                <div className="mb-5 flex space-x-4">
-                  <ButtonGroup color="secondary" size="medium">
-                    {Array(12)
-                      .fill(0)
-                      .map((_, i) => (
-                        <Button
-                          key={i}
-                          className={clsx({
-                            'brightness-125':
-                              state.activeBlock?.props.columns === i + 1,
-                          })}
-                          onClick={() => handleChangeColumnCount(i + 1)}
-                        >
-                          {i + 1}
-                        </Button>
-                      ))}
-                  </ButtonGroup>
+            const customTemplateColumns = splitColumns(value).reduce(
+              (a, b) => a + b,
+              0,
+            );
+
+            const customTemplateError = customTemplateColumns > 12;
+
+            return (
+              <div>
+                <p className="text-lg">Template</p>
+
+                <div className="flex flex-wrap">
+                  {templates.map((template) => (
+                    <TemplateButton
+                      key={template}
+                      template={template}
+                      onClick={handleChangeTemplate}
+                    />
+                  ))}
                 </div>
-              </div>,
-              view.propertyConfiguration.element,
+
+                <div className="py-4 text-center text-sm">
+                  or, Custom Columns
+                </div>
+
+                <div className="flex items-start">
+                  <div>
+                    <div>
+                      <Input
+                        placeholder="Template: 4+4+4"
+                        className={clsx(
+                          'mb-2',
+                          customTemplateError && 'border-red-500',
+                        )}
+                        value={value}
+                        onChange={(event) =>
+                          setValue(event.currentTarget.value)
+                        }
+                      />
+
+                      <Button
+                        className="mx-4"
+                        color="secondary"
+                        onClick={() =>
+                          handleChangeTemplate(splitColumns(value))
+                        }
+                      >
+                        Generate
+                      </Button>
+                    </div>
+
+                    {customTemplateError ? (
+                      <div className="text-sm text-red-500">
+                        The amount should not exceed 12, now{' '}
+                        {customTemplateColumns}
+                      </div>
+                    ) : (
+                      customTemplateColumns > 0 && (
+                        <div className="text-sm">
+                          {customTemplateColumns} columns out of 12
+                        </div>
+                      )
+                    )}
+                  </div>
+
+                  {!customTemplateError && (
+                    <TemplateButton
+                      className="p-0"
+                      template={value}
+                      onClick={handleChangeTemplate}
+                    />
+                  )}
+                </div>
+              </div>
             );
           },
         }),