+++ packages/uikit/src/index.ts
@@ -2,3 +2,4 @@ export * from './components/switch';
 export * from './components/button';
 export * from './components/button-group';
 export * from './components/icon-button';
+export * from './components/input';