+++ packages/react-editor/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @chamaeleon/react-editor
 
+## 0.0.7
+
+### Patch Changes
+
+- Renamed `plugin.view.update` to `plugin.view.render`
+- Updated dependencies
+  - @chamaeleon/uikit@0.1.0
+  - @chamaeleon/core@0.4.0
+
 ## 0.0.6
 
 ### Patch Changes